UI/requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appengine-python-standard==1.1.6
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
configparser==7.0.0
deprecation==2.1.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
Flask==3.0.3
Flask-Cors==4.0.1
frozendict==2.4.4
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mock==5.1.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.2
pydantic_core==2.18.3
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.2
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.35.0
streamlit-feedback==0.1.3
streamlit-modal==0.1.2
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.30.0
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
fuzzywuzzy==0.18.0
plotly-express==0.4.1
google-cloud-aiplatform==1.57.0
db-dtypes==1.2.0
st-pages
langchain-google-vertexai
st-pages==0.5.0
looker-sdk==24.12.0
statsmodels
looker-sdk